requirements.txt
alabaster==0.7.12
aspell-python-py3==1.15
Babel==2.10.3
beautifulsoup4==4.11.1
bracex==2.3.post1
CacheMan==2.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
dateparser==0.7.2
Django==2.2.4
docutils==0.19
future==0.18.2
hieroglyph==2.1.0
html5lib==1.1
humanize==0.5.1
idna==3.3
imagesize==1.4.1
Jinja2==3.1.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==21.3
pendulum==2.0.5
psutil==5.6.7
Pygments==2.13.0
pyparsing==3.0.9
pyspelling==2.8.1
python-dateutil==2.8.1
pytz==2019.3
pytzdata==2019.3
PyYAML==6.0
regex==2019.12.17
requests==2.28.1
six==1.13.0
snaptime==0.2.4
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.26.11
wcmatch==8.4
webencodings==0.5.1
